requirements.txt
pytest==8.3.2
pytest-base-url==2.1.0
requests==2.32.3
python-dotenv==1.0.1
pydantic~=2.9.2